End() should affect only one layer of Criteria
As part of the Silverstripe CMS 4 upgrade the Critera class was altered
to recursively end() all parent groups and return the FormField to which
the root Criteria was bound to.

This prevents `group()->rules()->end()->moreRules()->end()` from being a thing.
In particular, it prevents a Criteria from having multiple group rules.

The docblock states that end() will return either the FormField or
parent Criteria object - currently it only returns FormField which is
within that set, so I'm not sure if this constitutes a major change or
not.