Add Critera api to exit a single group() level
It was decided that 306f813 was too risky a change that could be
considered API breaking and thus a major version jump. To reduce this we
can add a new method for explicitly escaping a single group level
instead of making Criteria::end() do two different things depending on
context.

This adds the ability for authors to add multiple groups to their
display Criteria, without interrupting people with nested group() rules
that rely on end() to return the form field the rules apply to.